separated iex capella module
