+++ src/importexport/capella/CMakeLists.txt
@@ -1,0 +1,38 @@
+#=============================================================================
+#  MuseScore
+#  Music Composition & Notation
+#
+#  Copyright (C) 2020 MuseScore BVBA and others
+#
+#  This program is free software; you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License version 2.
+#
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with this program; if not, write to the Free Software
+#  Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
+#=============================================================================
+
+set(MODULE iex_capella)
+
+set(MODULE_SRC
+    ${CMAKE_CURRENT_LIST_DIR}/capellamodule.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/capellamodule.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/capellareader.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/capellareader.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/capella.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/capella.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/capxml.cpp
+    )
+
+set(MODULE_LINK
+    rtf2html    # for capella
+    libmscore
+    )
+
+include(${PROJECT_SOURCE_DIR}/build/module.cmake)
+
