--- src/importexport/internal/capella/capella.cmake
@@ -1,7 +1,0 @@
-
-
-set (CAPELLA_SRC
-    ${CMAKE_CURRENT_LIST_DIR}/capella.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/capella.h
-    ${CMAKE_CURRENT_LIST_DIR}/capxml.cpp
-    )
